formalize(endgame): WIP feat: sum_dist_diff_le (#124)
Joint work with @Paul-Lez. This PR
- completes the proof of the independence condition appearing in `sum_dist_diff_le` and
- adds a new lemma on joint independence in the Independence file. 

The remaining sorrys in `sum_dist_diff_le` are analogous to `ineq1` which is now formalized. I'd be very happy about any feedback on how to shorten the proofs of `ineq1`, `independenceCondition1` and `independenceCondition1'`. If these proofs can be simplified this would likely shorten the proofs of `ineq2`,..., `ineq6` as well.

Lean: sum_dist_diff_le
Status: math.open